6711 move attributes out of file search class
