+++ Core/src/org/sleuthkit/autopsy/discovery/search/Bundle.properties-MERGED
@@ -1,5 +1,3 @@
-AttributeSearchData.AttributeType.Domain.displayName=Domain
-AttributeSearchData.AttributeType.Other.displayName=Other
 DiscoveryKeyUtils.FileTagGroupKey.noSets=None
 DiscoveryKeyUtils.NoGroupingGroupKey.allFiles=All Files
 FileGroup.groupSortingAlgorithm.groupName.text=Group Name
@@ -26,39 +24,6 @@ FileSearch.HashHitsGroupKey.noHashHits=None
 FileSearch.InterestingItemGroupKey.noSets=None
 FileSearch.KeywordListGroupKey.noKeywords=None
 FileSearch.ObjectDetectedGroupKey.noSets=None
-FileSearchData.FileSize.100kbto1mb=: 100KB-1MB
-FileSearchData.FileSize.100mbto1gb=: 100MB-1GB
-FileSearchData.FileSize.10PlusGb=: 10GB+
-FileSearchData.FileSize.16kbto100kb=: 16-100KB
-FileSearchData.FileSize.1gbto5gb=: 1-5GB
-FileSearchData.FileSize.1mbto50mb=: 1-50MB
-FileSearchData.FileSize.200PlusMb=: 200MB+
-FileSearchData.FileSize.500kbto100mb=: 500KB-100MB
-FileSearchData.FileSize.50mbto200mb=: 50-200MB
-FileSearchData.FileSize.5gbto10gb=: 5-10GB
-FileSearchData.FileSize.LARGE.displayName=Large
-FileSearchData.FileSize.MEDIUM.displayName=Medium
-FileSearchData.FileSize.SMALL.displayName=Small
-FileSearchData.FileSize.upTo16kb=: 0-16KB
-FileSearchData.FileSize.upTo500kb=: 0-500KB
-FileSearchData.FileSize.XLARGE.displayName=XLarge
-FileSearchData.FileSize.XSMALL.displayName=XSmall
-FileSearchData.FileSize.XXLARGE.displayName=XXLarge
-FileSearchData.FileType.Audio.displayName=Audio
-FileSearchData.FileType.Documents.displayName=Documents
-FileSearchData.FileType.Executables.displayName=Executables
-FileSearchData.FileType.Image.displayName=Image
-FileSearchData.FileType.Other.displayName=Other/Unknown
-FileSearchData.FileType.Video.displayName=Video
-FileSearchData.Frequency.common.displayName=Common (11 - 100)
-FileSearchData.Frequency.known.displayName=Known (NSRL)
-FileSearchData.Frequency.rare.displayName=Rare (2-10)
-FileSearchData.Frequency.unique.displayName=Unique (1)
-FileSearchData.Frequency.unknown.displayName=Unknown
-FileSearchData.Frequency.verycommon.displayName=Very Common (100+)
-FileSearchData.Score.interesting.displayName=Interesting
-FileSearchData.Score.notable.displayName=Notable
-FileSearchData.Score.unknown.displayName=Unknown
 FileSearchFiltering.concatenateSetNamesForDisplay.comma=, 
 # {0} - filters
 FileSearchFiltering.HashSetFilter.desc=Hash set hits in set(s): {0}
@@ -79,6 +44,40 @@ ResultFile.score.interestingResult.description=At least one instance of the file
 ResultFile.score.notableFile.description=At least one instance of the file was recognized as notable.
 ResultFile.score.notableTaggedFile.description=At least one instance of the file is tagged with a notable tag.
 ResultFile.score.taggedFile.description=At least one instance of the file has been tagged.
+SearchData.AttributeType.Domain.displayName=Domains
+SearchData.FileSize.100kbto1mb=: 100KB-1MB
+SearchData.FileSize.100mbto1gb=: 100MB-1GB
+SearchData.FileSize.10PlusGb=: 10GB+
+SearchData.FileSize.16kbto100kb=: 16-100KB
+SearchData.FileSize.1gbto5gb=: 1-5GB
+SearchData.FileSize.1mbto50mb=: 1-50MB
+SearchData.FileSize.200PlusMb=: 200MB+
+SearchData.FileSize.500kbto100mb=: 500KB-100MB
+SearchData.FileSize.50mbto200mb=: 50-200MB
+SearchData.FileSize.5gbto10gb=: 5-10GB
+SearchData.FileSize.LARGE.displayName=Large
+SearchData.FileSize.MEDIUM.displayName=Medium
+SearchData.FileSize.SMALL.displayName=Small
+SearchData.FileSize.upTo16kb=: 0-16KB
+SearchData.FileSize.upTo500kb=: 0-500KB
+SearchData.FileSize.XLARGE.displayName=XLarge
+SearchData.FileSize.XSMALL.displayName=XSmall
+SearchData.FileSize.XXLARGE.displayName=XXLarge
+SearchData.FileType.Audio.displayName=Audio
+SearchData.FileType.Documents.displayName=Documents
+SearchData.FileType.Executables.displayName=Executables
+SearchData.FileType.Image.displayName=Image
+SearchData.FileType.Other.displayName=Other/Unknown
+SearchData.FileType.Video.displayName=Video
+SearchData.Frequency.common.displayName=Common (11 - 100)
+SearchData.Frequency.known.displayName=Known (NSRL)
+SearchData.Frequency.rare.displayName=Rare (2-10)
+SearchData.Frequency.unique.displayName=Unique (1)
+SearchData.Frequency.unknown.displayName=Unknown
+SearchData.Frequency.verycommon.displayName=Very Common (100+)
+SearchData.Score.interesting.displayName=Interesting
+SearchData.Score.notable.displayName=Notable
+SearchData.Score.unknown.displayName=Unknown
 # {0} - Data source name
 # {1} - Data source ID
 SearchFiltering.DataSourceFilter.datasource={0}({1})
